+++ react-container/src/components/index.ts
@@ -1,2 +1,3 @@
-export * from './CdfAuthContainer';
 export * from './AuthContainer';
+export * from './CdfAuthContainer';
+export * from './TranslationWrapper';