+++ react-container/src/utils/cluster.ts
@@ -1,4 +1,0 @@
-// this will end up from coming from sidecar eventually
-export const getClusterFromCdfApiBaseUrl = (cdfApiBaseUrl: string) => {
-  return cdfApiBaseUrl.split('https://')[1].split('.')[0];
-};