+++ react-tenant-selector/src/components/index.ts
@@ -5,3 +5,4 @@ export { default as CogniteFlow } from './CogniteFlow';
 export { default as LoginTip } from './LoginTip';
 export { default as TenantSelectorBackground } from './TenantSelectorBackground';
 export { default as TitleChanger } from './TitleChanger';
+export { default as ProjectSelector } from './ProjectSelector';