+++ react-tenant-selector/src/utils/sidecar.ts
@@ -1,22 +1,24 @@
 type SidecarConfig = {
-  appsApiBaseUrl: string;
-  cdfApiBaseUrl: string;
-  applicationId: string;
-  applicationName: string;
-  backgroundImage: string;
-  helpLink: string;
   AADClientID?: string;
   AADTenantID?: string;
+  applicationId: string;
+  applicationName: string;
+  appsApiBaseUrl: string;
+  backgroundImage: string;
+  cdfApiBaseUrl: string;
+  cdfCluster: string;
+  helpLink: string;
+  locizeProjectId?: string;
 };
 
 const DEFAULT_SIDECAR: SidecarConfig = {
   // -- Bluefield -- (Azure)
-  // appsApiBaseUrl: 'https://apps-api.bluefield.cognite.ai',
-  // cdfApiBaseUrl: 'https://bluefield.cognitedata.com',
+  appsApiBaseUrl: 'https://apps-api.bluefield.cognite.ai',
+  cdfApiBaseUrl: 'https://bluefield.cognitedata.com',
 
   // -- EW1 --
-  appsApiBaseUrl: 'https://apps-api.staging.cognite.ai',
-  cdfApiBaseUrl: 'https://api.cognitedata.com',
+  // appsApiBaseUrl: 'https://apps-api.staging.cognite.ai',
+  // cdfApiBaseUrl: 'https://api.cognitedata.com',
 
   // -- Greenfield --
   // appsApiBaseUrl: 'https://apps-api.greenfield.cognite.ai',
@@ -27,9 +29,20 @@ const DEFAULT_SIDECAR: SidecarConfig = {
   backgroundImage: '',
   helpLink: '',
 
+  // cdfCluster: 'greenfield',
+  cdfCluster: 'bluefield',
+  // cdfCluster: 'ew1',
+
+  // for localhost testing:
+  // locizeProjectId: 'dfcacf1f-a7aa-4cc2-94d7-de6ea4e66f1d', // tenant-selector project
+
   // use the demo app if you are testing this:
-  // AADClientID: 'b7c57241-915b-495a-bd0c-f13140f1c93b', // Cognite React Demo APP: (staging)
-  // AADTenantID: 'a9ae5b54-3600-4917-a9dc-3020723360b3', // cognitedata
+  AADTenantID: 'reactdemoapp.onmicrosoft.com',
+  // AADTenantID: `cognitedata.com`,
+
+  AADClientID: '7e8b0cb9-c1bf-46df-bf79-13cb1b19a8f8', // Demo App: bluefield (staging)
+  // AADClientID: '52106ea0-e93b-40bc-99dc-5f7deacceda1', // Demo App: ew1 (staging)
+  // AADClientID: 'e0e9ac77-f9cc-49ca-9bb2-eef01313751b', // Demo App: greenfield (staging)
 };
 
 export default (): SidecarConfig => {