+++ i18n/CHANGELOG.md
@@ -1,11 +1,17 @@
-# 0.2.3 - Jan 25 2020
+## 0.2.4 - Feb 5 2021
+
+### Changes
+
+Feat: return i18n instance after configure is complete
+
+## 0.2.3 - Jan 25 2021
 
 Feat: add option `localStorageLanguageKey`
 
-# 0.2.2
+## 0.2.2
 
 Chore: typescript type fixes
 
-# 0.2.1
+## 0.2.1
 
 Feat: bumped libs - "i18next": "^19.8.4", - "react-i18next": "^11.8.5",